Initial import
This copies in the postgresql server role we've used internally for
years now, with a few modifications to modernize it and prepare it for
public release.

The biggest difference between this role and others is that we
exclusively use PKI certificate-based authentication for clients, and
intend to support more enterprise platform targets.